requirements.txt
wheel==0.22.0
django-oauth2-provider